compatibility: seperate helpers into files
Refactoring to support multiple Kafka clients:
* Factor out helper classes into their own file.
* Use abstract factory pattern for auto-creating each client helper
  factory.
